pyproject.toml
[project]
name = "crypto"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.5.30",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.6",
    "mplfinance>=0.12.10b0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
